Permissions UI with user requests and modification post-saving.
Commits:

Basic, functional permission tab when requesting getPubKey. Only allows one time deny.

getPubKey and getRelays for sure working after prompt.

Changed the prompt to use a query string instead of background script queue. This should prevent any disconnect between the user expecting one thing and getting a different prompt.

It is not working using query string and working quite nicely.

Finally figured out the secret sauce to only open one prompt at a time.

Allow and deny now both work, with the option to remember the request next time.

Still tweaking lots of events to try and get the prompts working smoothly

Nice rendering for the event query message.

Tweaking the migration setup and the tab opening/closing code.

When remembering signing events, it is now scoped by event kind, as well!

Include extra event information in event signing dialogs.

Change confirm buttons to confirm dialog box.

Update nostr-tools to 1.2.1

The interface for app permissions looks good. Ready to work on functionality now.

Don't show app permissions section unless there are things to show.

Fix bug when saving a "Deny".

Additional formatting changes to options page for permissions UI.

Fine permissions seem to be working nicely!

Quick usability fix so that App Permissions section appears on Options page, even when there are no options selected.

Bumping build to #5. Preparing for new build release.

+++ Shared (Extension)/Resources/options.js
@@ -11,12 +11,16 @@ import {
     saveProfileName,
     saveRelays,
     RECOMMENDED_RELAYS,
+    getPermissions,
+    setPermission,
+    KINDS,
+    humanPermission,
 } from './utils';
 
 const log = console.log;
 
 Alpine.data('options', () => ({
-    profileNames: ['Poop'],
+    profileNames: ['---'],
     profileIndex: 0,
     profileName: '',
     pristineProfileName: '',
@@ -27,8 +31,11 @@ Alpine.data('options', () => ({
     newRelay: '',
     urlError: '',
     recommendedRelay: '',
-    confirmDelete: false,
-    confirmClear: false,
+    permissions: {},
+    host: '',
+    permHosts: [],
+    hostPerms: [],
+    setPermission,
 
     async init(watch = true) {
         log('Initialize backend.');
@@ -37,6 +44,11 @@ Alpine.data('options', () => ({
         if (watch) {
             this.$watch('profileIndex', async () => {
                 await this.refreshProfile();
+                this.host = '';
+            });
+
+            this.$watch('host', () => {
+                this.calcHostPerms();
             });
 
             this.$watch('recommendedRelay', async () => {
@@ -59,8 +71,7 @@ Alpine.data('options', () => ({
         await this.getNsec();
         await this.getNpub();
         await this.getRelays();
-        this.confirmClear = false;
-        this.confirmDelete = false;
+        await this.getPermissions();
     },
 
     // Profile functions
@@ -87,8 +98,14 @@ Alpine.data('options', () => ({
     },
 
     async deleteProfile() {
-        await deleteProfile(this.profileIndex);
-        await this.init(false);
+        if (
+            confirm(
+                'This will delete this profile and all associated data. Are you sure you wish to continue?'
+            )
+        ) {
+            await deleteProfile(this.profileIndex);
+            await this.init(false);
+        }
     },
 
     // Key functions
@@ -96,9 +113,13 @@ Alpine.data('options', () => ({
     async saveProfile() {
         if (!this.needsSave) return;
 
+        console.log('saving private key');
         await savePrivateKey(this.profileIndex, this.privKey);
+        console.log('saving profile name');
         await saveProfileName(this.profileIndex, this.profileName);
+        console.log('getting profile name');
         await this.getProfileNames();
+        console.log('refreshing profile');
         await this.refreshProfile();
     },
 
@@ -161,11 +182,57 @@ Alpine.data('options', () => ({
         }, 3000);
     },
 
+    // Permissions
+
+    async getPermissions() {
+        this.permissions = await getPermissions(this.profileIndex);
+
+        // Set the convenience variables
+        this.calcPermHosts();
+        this.calcHostPerms();
+    },
+
+    calcPermHosts() {
+        let hosts = Object.keys(this.permissions);
+        hosts.sort();
+        this.permHosts = hosts;
+    },
+
+    calcHostPerms() {
+        let hp = this.permissions[this.host] || {};
+        let keys = Object.keys(hp);
+        keys.sort();
+        this.hostPerms = keys.map(k => [k, humanPermission(k), hp[k]]);
+        console.log(this.hostPerms);
+    },
+
+    permTypes(hostPerms) {
+        let k = Object.keys(hostPerms);
+        k = Object.keys.sort();
+        k = k.map(p => {
+            let e = [p, hostPerms[p]];
+            if (p.startsWith('signEvent')) {
+                let n = parseInt(p.split(':')[1]);
+                let name =
+                    KINDS.find(kind => kind[0] === n) || `Unknown (Kind ${n})`;
+                e = [name, hostPerms[p]];
+            }
+            return e;
+        });
+        return k;
+    },
+
     // General
 
     async clearData() {
-        await clearData();
-        await this.init(false);
+        if (
+            confirm(
+                'This will remove your private keys and all associated data. Are you sure you wish to continue?'
+            )
+        ) {
+            await clearData();
+            await this.init(false);
+        }
     },
 
     // Properties
